ai_marketing_agent/requirements.txt
google-generativeai
PyYAML
